+++ src/pro/blocks/woo-comparison/index.js
@@ -3,67 +3,75 @@
  */
 import { __ } from '@wordpress/i18n';
 
-import { registerBlockType } from '@wordpress/blocks';
+import {
+	createBlock,
+	getBlockTypes,
+	registerBlockType
+} from '@wordpress/blocks';
 
 import { useBlockProps } from '@wordpress/block-editor';
 
-import { Placeholder } from '@wordpress/components';
+import {
+	Button,
+	Placeholder
+} from '@wordpress/components';
+
+import { useDispatch } from '@wordpress/data';
+
+import { useEffect } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
 import metadata from './block.json';
-import attributes from './attributes.js';
-import edit from './edit.js';
-import Inactive from '../../components/inactive/index.js';
-
-const { faIcon: icon } = window.otterUtils.icons;
 
 const { name } = metadata;
 
-if ( Boolean( window.otterPro.hasWooCommerce ) && Boolean( window.otterPro.hasNeveSupport.wooComparison ) ) {
-	if ( ! ( Boolean( window.otterPro.isActive ) && ! Boolean( window.otterPro.isExpired ) ) ) {
-		edit = () => <Inactive
-			icon={ icon }
-			label={ metadata.title }
-			blockProps={ useBlockProps() }
-		/>;
+let shouldMigrate = false;
+
+const edit = ({
+	clientId,
+	attributes
+}) => {
+	const { replaceBlocks } = useDispatch( 'core/block-editor' );
+
+	useEffect( () => {
+		const blocks = getBlockTypes();
+
+		if ( -1 !== blocks.findIndex( block => 'sparks/woo-comparison' === block.name ) ) {
+			shouldMigrate = true;
+		}
+	}, []);
+
+	if ( shouldMigrate ) {
+		replaceBlocks( clientId, [
+			createBlock( 'sparks/woo-comparison', {
+				...attributes
+			})
+		]);
 	}
 
-	registerBlockType( name, {
-		...metadata,
-		title: __( 'WooCommerce Comparison Table', 'otter-blocks' ),
-		description: __( 'A way to compare different WooCommerce products made on the website. Powered by Otter.', 'otter-blocks' ),
-		icon,
-		keywords: [
-			'woocommerce',
-			'comparison',
-			'table'
-		],
-		attributes,
-		supports: {
-			inserter: Boolean( window.otterPro.isActive ) && ! Boolean( window.otterPro.isExpired )
-		},
-		edit,
-		save: () => null
-	});
-} else {
-	registerBlockType( name, {
-		...metadata,
-		title: __( 'WooCommerce Comparison Table', 'otter-blocks' ),
-		description: __( 'A way to compare different WooCommerce products made on the website. Powered by Otter.', 'otter-blocks' ),
-		icon,
-		keywords: [
-			'woocommerce',
-			'comparison',
-			'table'
-		],
-		attributes,
-		supports: {
-			inserter: false
-		},
-		edit: () => <div { ...useBlockProps() }><Placeholder>{ __( 'You need to have Neve Pro & WooCommerce installed to edit WooCommerce Comparison Table block.', 'otter-blocks' ) }</Placeholder></div>,
-		save: () => null
-	});
-}
+	return (
+		<div { ...useBlockProps() }>
+			<Placeholder>{ __( 'You need to install the latest version of Neve with Sparks for WooCommerce to use WooCommerce Comparison Table.', 'otter-blocks' ) }</Placeholder>
+		</div>
+	);
+};
+
+registerBlockType( name, {
+	...metadata,
+	title: __( 'WooCommerce Comparison Table', 'otter-blocks' ),
+	description: __( 'A way to compare different WooCommerce products made on the website.', 'otter-blocks' ),
+	icon: 'editor-table',
+	keywords: [
+		'woocommerce',
+		'comparison',
+		'table'
+	],
+	supports: {
+		inserter: false
+	},
+	edit,
+	save: () => null
+});